refactor(auth): organize login method flows
